Refactor - click handler 각각 파일로 분리
